Move report interpretation helpers to shared boundaries
Relocate the report-specific interpretation helpers and PrimaryReportFacts out of use_cases.history into shared/boundaries, update history and PDF consumers to the shared seam, and add a focused contract assertion that report preparation resolves those facts from the shared module.
