+++ apps/server/tests/use_cases/history/test_report_preparation_contract.py
@@ -7,8 +7,10 @@ from test_support.findings import make_finding_payload
 
 from vibesensor.adapters.pdf import mapping as pdf_mapping
 from vibesensor.adapters.pdf.report_context import prepare_report_mapping_context
+from vibesensor.shared.boundaries import report_interpretation as shared_report_interpretation
 from vibesensor.use_cases.history.report_preparation import (
     PreparedReportInput,
+    PrimaryReportFacts,
     ValidatedPreparedReportInput,
     prepare_report_input,
     validate_prepared_report_input,
@@ -91,3 +93,7 @@ def test_map_summary_fails_before_pdf_mapping_for_invalid_prepared_input(
 
     with pytest.raises(ValueError, match="report_facts"):
         pdf_mapping.map_summary(prepared)
+
+
+def test_report_preparation_imports_primary_report_facts_from_shared_boundaries() -> None:
+    assert PrimaryReportFacts is shared_report_interpretation.PrimaryReportFacts
