+++ apps/server/vibesensor/adapters/pdf/_candidate_resolver.py
@@ -12,7 +12,7 @@ from vibesensor.report_i18n import human_source
 
 if TYPE_CHECKING:
     from vibesensor.adapters.pdf.report_context import ReportMappingContext
-    from vibesensor.use_cases.history.report_interpretation import PrimaryReportFacts
+    from vibesensor.shared.boundaries.report_interpretation import PrimaryReportFacts
 
 __all__ = [
     "PrimaryCandidateContext",
