+++ apps/server/vibesensor/use_cases/history/report_preparation.py
@@ -21,15 +21,7 @@ from vibesensor.shared.boundaries.analysis_payload import (
     SummaryWarningPayload,
 )
 from vibesensor.shared.boundaries.diagnostic_case import test_run_from_summary
-from vibesensor.shared.boundaries.run_suitability import run_suitability_payload
-from vibesensor.shared.boundaries.summary_warning import summary_warning_payloads
-from vibesensor.shared.json_utils import as_float_or_none as _as_float
-from vibesensor.shared.time_utils import utc_now_iso
-from vibesensor.shared.types.history_analysis_contracts import AnalysisSummary
-from vibesensor.shared.types.persisted_analysis import PersistedAnalysis
-from vibesensor.use_cases.history.helpers import safe_filename
-from vibesensor.use_cases.history.report_cache import ReportPdfCacheKey
-from vibesensor.use_cases.history.report_interpretation import (
+from vibesensor.shared.boundaries.report_interpretation import (
     PrimaryReportFacts,
     compute_location_hotspot_rows,
     filter_active_sensor_intensity,
@@ -38,6 +30,14 @@ from vibesensor.use_cases.history.report_interpretation import (
     resolve_report_origin,
     tire_spec_text,
 )
+from vibesensor.shared.boundaries.run_suitability import run_suitability_payload
+from vibesensor.shared.boundaries.summary_warning import summary_warning_payloads
+from vibesensor.shared.json_utils import as_float_or_none as _as_float
+from vibesensor.shared.time_utils import utc_now_iso
+from vibesensor.shared.types.history_analysis_contracts import AnalysisSummary
+from vibesensor.shared.types.persisted_analysis import PersistedAnalysis
+from vibesensor.use_cases.history.helpers import safe_filename
+from vibesensor.use_cases.history.report_cache import ReportPdfCacheKey
 
 if TYPE_CHECKING:
     from vibesensor.domain import Finding, TestRun
